make security gallery policy permission

+++ app/Http/Controllers/Admin/GalleryController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\GalleryRequest;
+use App\Models\Gallery;
 use App\Services\GalleryService;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Log;
 
 class GalleryController extends Controller
@@ -22,6 +24,9 @@ class GalleryController extends Controller
 
     public function store(GalleryRequest $request): RedirectResponse
     {
+        // check permission
+        Gate::authorize('create', Gallery::class);
+
         $credentials = $request->validated();
 
         if ($request->file('image')) {
@@ -38,6 +43,10 @@ class GalleryController extends Controller
 
     public function destroy(int $id): RedirectResponse
     {
+        // check permission
+        $gallery = Gallery::where('id', $id)->firstOrFail();
+        Gate::authorize('delete', $gallery);
+
         DB::transaction(function () use ($id) {
             $gallery = $this->galleryService->delete($id);
             Log::info($gallery);
